requirements.txt
# flask
# pymongo[srv]
# python-dotenv
# gunicorn


Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==0.21.1
Werkzeug==2.2.3